サイクル一覧ページと消去済みページで、カードのコンポーネントを使い回すように改修
