--- front/react_app/src/pages/home/CycleCard.tsx
@@ -1,94 +1,0 @@
-import {
-  faArrowsSpin,
-  faEllipsisVertical,
-} from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import React from "react";
-import { CycleIfc } from "../../utils/cycle.interface";
-import { Link } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import { showModal } from "../../redux/modalSlice";
-import { hideMenu, showMenu } from "../../redux/menuSlice";
-import { RootState } from "../../redux/store";
-
-interface CycleInfoProps {
-  element: CycleIfc;
-}
-
-const CycleCard = (props: CycleInfoProps) => {
-  const menuState = useSelector((state: RootState) => state.menu.value);
-  const dispatch = useDispatch();
-
-  const MenuHandler = (e: React.FormEvent, cycleId: number) => {
-    e.preventDefault();
-    if (menuState.visible) {
-      dispatch(hideMenu());
-    } else {
-      dispatch(showMenu(cycleId));
-    }
-  };
-
-  const FavoriteHandler = (e: React.FormEvent, cycleId: number) => {
-    e.preventDefault();
-    dispatch(hideMenu());
-    console.log("Favorite!");
-    console.log(`CycleId: ${cycleId}`);
-  };
-
-  const ElaseHandler = (
-    e: React.FormEvent,
-    cycleId: number,
-    cycleName: string,
-  ) => {
-    e.preventDefault();
-    dispatch(hideMenu());
-    dispatch(showModal({ cycleId, cycleName }));
-  };
-
-  return (
-    <div>
-      <Link to={`/${props.element.id}`}>
-        <div className="flex h-28 transition duration-400 hover:bg-slate-200 hover:cursor-pointer select-none">
-          <div className="w-1/6 flex items-center justify-center">
-            <FontAwesomeIcon
-              icon={faArrowsSpin}
-              className="text-5xl text-blue-600"
-            />
-          </div>
-          <div className="w-5/6 pt-2 border-b border-slate-100 flex justify-between">
-            <h2 className="text-lg">{props.element.name}</h2>
-            <div className="flex justify-between">
-              {menuState.visible && menuState.cycleId === props.element.id && (
-                <div>
-                  <div className="mt-3 mr-3 menu">
-                    <p
-                      onClick={(e) => FavoriteHandler(e, props.element.id)}
-                      className="bg-slate-300 hover:bg-slate-400 px-3 py-1 rounded-t-md"
-                    >
-                      お気に入りに登録
-                    </p>
-                    <p
-                      onClick={(e) =>
-                        ElaseHandler(e, props.element.id, props.element.name)
-                      }
-                      className="bg-slate-300 hover:bg-slate-400 px-3 py-1 rounded-b-md"
-                    >
-                      消去する
-                    </p>
-                  </div>
-                </div>
-              )}
-              <FontAwesomeIcon
-                icon={faEllipsisVertical}
-                className="hover:bg-slate-300 px-3 py-2 mr-3 text-xl rounded transition duration-400"
-                onClick={(e) => MenuHandler(e, props.element.id)}
-              />
-            </div>
-          </div>
-        </div>
-      </Link>
-    </div>
-  );
-};
-
-export default CycleCard;
