+++ front/react_app/src/pages/home/Home.tsx
@@ -2,7 +2,7 @@ import { faCircleNotch } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import axios from "axios";
 import { useEffect, useState } from "react";
-import CycleCard from "./CycleCard";
+import CycleCard from "./HomeCycleCard";
 import Layout from "../../components/Layout";
 import { CycleIfc } from "../../utils/cycle.interface";
 import Header from "./Header";
